feat: creating a new layout and group to handle matches routes
